+++ hutool-core/src/test/java/cn/hutool/core/util/XmlUtilTest.java
@@ -1,5 +1,6 @@
 package cn.hutool.core.util;
 
+import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.io.resource.ResourceUtil;
 import cn.hutool.core.lang.Console;
@@ -212,6 +213,29 @@ public class XmlUtilTest {
 		Assert.assertEquals(testBean.getOperator(), testBean2.getOperator());
 		Assert.assertEquals(testBean.getProjectCode(), testBean2.getProjectCode());
 		Assert.assertEquals(testBean.getBankCode(), testBean2.getBankCode());
+	}
+
+	@Test
+	public void xmlToBeanTest2(){
+		//issue#1663@Github
+		String xmlStr = "<?xml version=\"1.0\" encoding=\"gbk\" ?><response><code>02</code></response>";
+
+		Document doc = XmlUtil.parseXml(xmlStr);
+
+		// 标准方式
+		Map<String, Object> map = XmlUtil.xmlToMap(doc.getFirstChild());
+		SmsRes res = new SmsRes();
+		BeanUtil.fillBeanWithMap(map, res, true);
+
+		// toBean方式
+		SmsRes res1 = XmlUtil.xmlToBean(doc.getFirstChild(), SmsRes.class);
+
+		Assert.assertEquals(res.toString(), res1.toString());
+	}
+
+	@Data
+	static class SmsRes {
+		private String code;
 	}
 
 	@Test